Parse column schema as data frame
This allows us to manipulate the columns specification in a more idiomatic way - e.g. instead of `lapply` we can use data frame accessors.

This is particularly useful for the 6th csvw test where we need to replace column names with `csvw:titles` when creating object keys in the list representation of the table.

This means the R representation of the metadata schema diverges from the tree structure of the json. It ought however be isomorphic. We can defer for now the writing of a translation (back into a list of list) for serialising to json.