Re-work table normalisation
This patch takes us further into metadata normalisation.

We establish a `normalise_property` method, that currently only affects link properties. These are now resolved against the base url. This in turn requires us to pass the location alongwith the metadata to the `normalise_metadata` function. We also implement a lookup to determine the `property_type` from the name of the annotation property. The implementation of URL normalisation with `base_url`, `resolve_url` and `normalise_url` is all quite naïve at this stage.

This patch also extends the complexity of tables are rendered as json.