Introduce dialect
Dialect is used in `add_dataframe` parsing and `table_to_list` transformation. A `default_dialect` from the spec is provided. The `default_schema` param in those functions is renamed to `group_schema` for clarity. A `default_schema` is introduced deriving function providing e.g. "_col.1" headings.

If neither table nor group schema is provided then the derived schema is attached to the table.

Import `rlang::%||%` for coalescing nulls with defaults.