Locate local tables
Introduces a tryCatch allowing `read_csvw` to succeed even when it wasn't possible to `add_dataframe`. A `dataframe` key is still added to the `table`, instead of a data frame it provides a list explaining the error.

The metadata pre-processing is split into `parse_metadata` which happens before adding a data frame and `normalise_metadata` which now happens afterwards. This is so that non-normalised URLs may be used when attempting to locate the data frame.