Workaround for fragment expansion when rendering URI templates
The URI template spec supports [fragment identifiers](https://datatracker.ietf.org/doc/html/rfc6570#section-3.2.4)
by expanding a template like `{#var}`. Glue interprets the hash char as a comment as it's inside curly braces.
[A recent change to glue](https://github.com/tidyverse/glue/issues/224) flushed this out.

This patch rewrites the template as `#{var}`. This has the desired effect when the value is present but doesn't
conform with the spec properly as the whole fragment (including the `#`) should be omitted. This is raised in #4.